Refactor the cmake compile options a bit (#129)
* Refactor the cmake compile options a bit

A few options (like -DPLATFORM=1 and macos filesystem and warnings)
were starting to be applied multiple places and not consistently.
For instance I had forgotten them on the auv2 build helper. So

1. Make a `clap-wrapper-compile-options` interface which sets up
   all these things
2. Set it as a link dep of the things which need it
3. Fix the build bugs with the auv2 helper from turning on warnings
   there and
4. Add an option to likn with asan, which is still a work in progress

* Split the sanitize from the warning and platform

* Respond to 0ax1 review comments

+++ src/detail/auv2/auv2_shared.h
@@ -53,7 +53,7 @@ struct ClapBridge
 
     if (_clapid.empty())
     {
-      if (_idx < 0 || _idx >= _library.plugins.size())
+      if (_idx < 0 || _idx >= (int)_library.plugins.size())
       {
         std::cout << "[ERROR] cannot load by index" << std::endl;
         return;
